+++ apps/demo-map/src/views/AllMapView.vue
@@ -11,6 +11,7 @@ import {
   MouseCoordinatesControl,
   SettingControl,
   ZoomControl,
+  langStore,
 } from '@hungpvq/vue-map-core';
 import {
   DrawControl,
@@ -30,6 +31,152 @@ import { MeasurementControl } from '@hungpvq/vue-map-measurement';
 import { PrintAdvancedControl, PrintControl } from '@hungpvq/vue-map-print';
 import AsideControl from '../layout/aside-control.vue';
 function onMapLoaded(map: MapSimple) {
+  langStore.setMapLang(map.id, {
+    map: {
+      basemap: {
+        title: 'Bản Đồ nền',
+        setting: 'Cài Đặt',
+      },
+      'crs-control': {
+        title: 'Cài đặt CRS',
+        field: {
+          name: 'tên',
+          unit: 'đơn vị',
+          epsg: 'EPSG',
+          proj4js: 'proj4js',
+        },
+      },
+      action: {
+        'fullscreen-control-enter': 'Toàn màn hình',
+        'fullscreen-control-exit': 'Thoát toàn màn hình',
+        'geolocate-control-find-my-location': 'Tìm vị trí của tôi',
+        'geolocate-control-location-not-available': 'Vị trí không có sẵn',
+        'navigation-control-zoom-in': 'Phóng to',
+        'navigation-control-zoom-out': 'Thu phóng ra',
+        'navigation-control-reset-bearing': 'Đặt lại về phía bắc',
+      },
+      'goto-control': {
+        title: 'Đi đến',
+        field: {
+          zoom: 'Phóng',
+          center: 'Trung tâm',
+          sprite: 'URL sprite',
+          glyphs: 'Url glyphs',
+        },
+        btn: {
+          apply: 'Đi đến',
+        },
+      },
+      home: {
+        title: 'Trang chủ',
+      },
+      'setting-control': {
+        title: 'Cài đặt',
+        field: {
+          zoom: 'Phóng',
+          center: 'Trung tâm',
+          sprite: 'URL sprite',
+          glyphs: 'Url glyphs',
+        },
+        btn: {
+          apply: 'Áp dụng',
+        },
+      },
+      'layer-control': {
+        title: 'Kiểm soát lớp',
+        'create-btn': 'Tạo lớp',
+        create: {
+          title: 'Lớp mới',
+        },
+        field: {
+          name: 'Tên',
+          type: 'Kiểu',
+          url: 'URL',
+          minzoom: 'Thu phóng tối thiểu',
+          maxzoom: 'Phóng to tối đa',
+          file: 'Tài liệu',
+          geojson: 'Geojson',
+          tiles: 'Gạch',
+          bound: {
+            title: 'Ràng buộc',
+            minx: 'Kinh độ tối thiểu',
+            miny: 'Vĩ độ tối thiểu',
+            maxx: 'Kinh độ tối đa',
+            maxy: 'Vĩ độ tối đa',
+          },
+        },
+        info: {
+          title: 'Thông tin',
+        },
+      },
+      identify: {
+        title: 'Nhận dạng',
+        point: 'Điểm',
+      },
+      measurement: {
+        action: {
+          clear: 'Thông thoáng',
+          close: 'Đóng',
+          setting: 'Cài đặt',
+          download: 'Tải xuống',
+          'add-point': 'Thêm điểm',
+          'fly-to': 'Lấp đầy ràng buộc',
+          add: 'Thêm vào',
+        },
+        title: 'Đo lường',
+        result: 'Kết quả đo lường',
+        field: {
+          'unit-distance': 'Khoảng cách đơn vị',
+          'unit-area': 'Khu vực đơn vị',
+        },
+        tools: {
+          point: 'Đo điểm',
+          distance: 'Đo khoảng cách',
+          area: 'Đo diện tích',
+          azimuth: 'Đo góc phương vị',
+        },
+        unit: {
+          meter: 'Mét',
+          kilometer: 'Km',
+          'square-meter': 'Mét vuông',
+          hecta: 'HECTA',
+          'square-kilometer': 'Km vuông',
+        },
+        setting: {
+          title: 'Cài đặt',
+          field: {
+            data: 'Dữ liệu',
+          },
+          point: 'Điểm',
+          distance: 'Khoảng cách',
+          area: 'Khu vực',
+          azimuth: 'Phương vị',
+        },
+        'no-data': {
+          text: 'Trạng thái',
+          value: 'Chờ...',
+        },
+      },
+      print: {
+        title: 'In',
+        actions: {
+          save: 'cứu',
+          clear: 'thông thoáng',
+          setting: 'Cài đặt',
+        },
+        setting: {
+          title: 'Cài đặt',
+        },
+        field: {
+          ratio: 'Tỷ lệ',
+          orientation: 'Định hướng',
+        },
+        btn: {
+          apply: 'In',
+        },
+      },
+    },
+  });
   addLayer(
     map.id,
     createRasterUrlLayer({