+++ libs/map/core/src/extra/lang/store.ts
@@ -6,6 +6,8 @@ import { addToQueue } from '../../store/queue';
 const KEY = 'lang';
 export type MapLocateStore = {
   locale: any;
+  localeDefault: any;
+  translate?: (key: string, params?: Record<string, any>) => string;
 };
 
 function initMapLang(mapId: string) {
@@ -14,19 +16,44 @@ function initMapLang(mapId: string) {
     KEY,
     reactive({
       locale: {},
+      localeDefault: {},
     })
   );
 }
+
 export function getMapLang(mapId: string) {
   const store = getStore<MapLocateStore>(mapId, KEY);
-  const storeLang = store.locale;
-  return storeLang;
+  if (!store) {
+    initMapLang(mapId);
+    return getStore<MapLocateStore>(mapId, KEY);
+  }
+  return store;
 }
 
-export function setMapLang(mapId: string, locale: Record<string, string>) {
+export function setMapLang(mapId: string, locale: Record<string, any>) {
   const store = getMapLang(mapId);
   if (store) {
-    store.locale = merge(store.locale || {}, locale);
+    store.locale = merge({}, store.locale, locale);
+  }
+}
+
+export function setMapLocaleDefault(
+  mapId: string,
+  locale: Record<string, any>
+) {
+  const store = getMapLang(mapId);
+  if (store) {
+    store.localeDefault = merge({}, store.localeDefault, locale);
+  }
+}
+
+export function setMapTranslate(
+  mapId: string,
+  translate: (key: string, params?: Record<string, any>) => string
+) {
+  const store = getMapLang(mapId);
+  if (store) {
+    store.translate = translate;
   }
 }