+++ libs/map/core/src/field/input-textarea.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-const model = defineModel();
+const model = defineModel<string>();
 defineProps({
   label: String,
 });