+++ libs/map/layer/src/modules/LayerControl.vue
@@ -36,8 +36,8 @@ const props = defineProps({
   disabled: Boolean,
 });
 const { mapId, moduleContainerProps } = useMap(props);
-const { trans, setLocale } = useLang(mapId.value);
-setLocale({
+const { trans, setLocaleDefault } = useLang(mapId.value);
+setLocaleDefault({
   map: {
     'layer-control': {
       title: 'Layer Control',