+++ libs/map/basemap/CHANGELOG.md
@@ -1,3 +1,25 @@
+## 0.2.0 (2025-04-14)
+
+### 🚀 Features
+
+- change locale
+
+### 🩹 Fixes
+
+- build
+
+### ❤️ Thank You
+
+- hung.pv
+
+## 0.1.0 (2025-04-14)
+
+### 🚀 Features
+
+- change locale
+
+## 0.0.3 (2025-04-14)
+
 ## 0.0.2 (2024-06-17)
 
 This was a version bump only for map-basemap to align it with other projects, there were no code changes.