+++ libs/router/CHANGELOG.md
@@ -1,3 +1,21 @@
+## 0.2.0 (2025-04-14)
+
+### 🚀 Features
+
+- init manga
+
+- **route:** create route
+
+### 🩹 Fixes
+
+- build
+
+### ❤️ Thank You
+
+- hung.pv
+
+## 0.1.1 (2025-04-14)
+
 ## 0.1.0 (2025-04-13)
 
 ### 🚀 Features
@@ -5,7 +23,3 @@
 - init manga ([b6a72d2](https://github.com/hung4564/vue-library/commit/b6a72d2))
 
 - **route:** create route ([25e643d](https://github.com/hung4564/vue-library/commit/25e643d))
-
-### ❤️ Thank You
-
-- hung.pv