+++ libs/share/core/CHANGELOG.md
@@ -1,12 +1,22 @@
+## 0.0.5 (2025-04-14)
+
+### 🩹 Fixes
+
+- add missing export
+
+- build
+
+### ❤️ Thank You
+
+- hung.pv
+
+## 0.0.4 (2025-04-14)
+
 ## 0.0.3 (2025-04-13)
 
 ### 🩹 Fixes
 
 - add missing export ([5bf23ad](https://github.com/hung4564/vue-library/commit/5bf23ad))
-
-### ❤️ Thank You
-
-- hung.pv
 
 ## 0.0.2 (2024-06-17)