+++ libs/share/map/CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.0.5 (2025-04-14)
+
+### 🩹 Fixes
+
+- build
+
+### ❤️ Thank You
+
+- hung.pv
+
+## 0.0.4 (2025-04-14)
+
+## 0.0.3 (2025-04-14)
+
 ## 0.0.2 (2024-06-17)
 
 This was a version bump only for shared-map to align it with other projects, there were no code changes.