+++ libs/share/shared/CHANGELOG.md
@@ -1,3 +1,25 @@
+## 0.2.0 (2025-04-14)
+
+### 🚀 Features
+
+- list, aside control
+
+- editor io custom
+
+- **shared:** add new store
+
+### 🩹 Fixes
+
+- doc
+
+- build
+
+### ❤️ Thank You
+
+- hung.pv
+
+## 0.1.1 (2025-04-14)
+
 ## 0.1.0 (2025-04-13)
 
 ### 🚀 Features
@@ -11,10 +33,6 @@
 ### 🩹 Fixes
 
 - doc ([6a97b81](https://github.com/hung4564/vue-library/commit/6a97b81))
-
-### ❤️ Thank You
-
-- hung.pv
 
 ## 0.0.2 (2024-06-17)