+++ libs/ui/core/CHANGELOG.md
@@ -1,3 +1,21 @@
+## 0.2.0 (2025-04-14)
+
+### 🚀 Features
+
+- init manga
+
+- list, aside control
+
+### 🩹 Fixes
+
+- build
+
+### ❤️ Thank You
+
+- hung.pv
+
+## 0.1.1 (2025-04-14)
+
 ## 0.1.0 (2025-04-13)
 
 ### 🚀 Features
@@ -5,7 +23,3 @@
 - init manga ([b6a72d2](https://github.com/hung4564/vue-library/commit/b6a72d2))
 
 - list, aside control ([93a607f](https://github.com/hung4564/vue-library/commit/93a607f))
-
-### ❤️ Thank You
-
-- hung.pv